move diagram interface into DiagramCard component
