+++ src/types.ts
@@ -12,12 +12,3 @@ export enum Disease {
     Laminopathy = "Laminopathy",
     SkeletalMyopathy = "Skeletal Myopathy",
 }
-export interface Diagram {
-    image: {
-        childImageSharp: {
-            gatsbyImageData: IGatsbyImageData;
-        };
-    };
-    title: string;
-    caption: string;
-}
